Split pkg target's ZIP stage to new target
This change is mainly being made for the sake of intergrating into Github Actions (see next commit!), which has its own zipping functionality in the form of artifacts. Attempting to upload a ZIP this way creates a zip-within-a-zip, which isn't ideal. Calling "make zip" should produce identical behavior to the old "make pkg".