Calculate options menu item count
The options menu item count is now calculated on-the-fly, which
will make it easier to add or remove items to the options menu.
As a side effect, option_items is also now a local-varaible, so
if other functions need the string data, they can get it.

I attempted to split this declaration into its own header, but
the result always threw errors due to duplicate declares. Maybe
some other time!