Change abstract-button conversion to use SDL_Keycode array
This changes the switch cases in keyboard_to_abstract_button()
to use an array of SDL_Keycode values to see what keys to convert
to what buttons. This will allow for keyboard remapping to be
possible, at the cost of slightly uglier function code.

The reason for changing to if statements is that C/C++ doesn't
allow you to use array members ("case keymap[0]") as cases in
switch statements. Understandably so, but a bit of a shame.