Add new bg_data parameter: "shape_advanced"
This will allow background effects to do things whenever a shape
is succesfully placed by the player and we advanced to the next one.
Nothing currently uses this yet, though.

This commit also fixes an oversight in bg_test in which it wasn't
using the struct definition in background.h.