[BGFX] Calculate munch rates once, restore 256x texture size
The main bottleneck in the original version of munching was the
fact it calculated "sin(song_tick/munch_rate_X)" three times for
every single pixel, 60+ times a second. Since these values are
consistent across every pixel, we can calculate it once before
the for-loop, going from 256*256*3 calcs to just 3.

Needless to say, this dramatically improves performance of this
particular BGFX, and allows the original resolution to be used.

Also, I mistakenly labelled the texture a "moire". Different thing!