[GH-Actions-win64] Static-link GCC DLL, include libwinpthread-1.dll
Open Manifold required two DLLs to run that weren't included in
releases of the game; these two DLLs (libwinpthread-1 and
libgg_s_seh-1) are part of GCC and should've been static-linked.

Linking libwinpthread as static has proven to be difficult, so for now
it's simply included alongside the SDL2 DLLs.

Thanks to xen for bringing this to my attention!