Implement a timer for character status
Previously, characters would simply hold on whatever state they
were currently in until the end of the player phase. Now each state
will last for a fixed amount of time before reverting to idle.

At the moment, the amount of time is fixed to 400 milliseconds, but
I'd like to set it to the actual length of 1 beat relative to the
song's BPM in the future.

Solves #13.