Implement toolbar for sandbox mode
This commit introduces a new "toolbar" to Sandbox. This bar will
allow for the addition of several new features to Sandbox without
having to bind an ungodly number of buttons or keys.

This initial implementation only has one option, "Change Color":
When hovering over this tool, pressing Up or Down will cycle the
color of the current shape. This effectively solves #11.

I've reserved six icons in the new sandbox icons texture, so expect
to see more commits in the near future that implement more things
(some ideas I have include exporting to JSON, shape morph, color morph,
changing the color of the background, and grid toggle.)

The controls in sandbox have been changed to account for this
bar: the X-plode button now places shapes, and Start will open
the toolbar.