Load snd_combo in load_default_sound_colleection to fix segfault
An oversight while implementing the internal "unload_sounds()" function
resulted in a bug where loading a level, then Sandbox, then Sandbox again.
The game would attempt to free the memory of the already-freed "snd_combo",
resulting in a segfault.

To fix this, load_default_sound_collection now attempts to load snd_combo
despite the only mode that uses this function, Sandbox, not actually needing
to load this particular sound.

As a side note: this bug was found while its creator was running the game
on a recently-purchased monitor, just as a silly test. Ouch!