Check bounds of user-made sequence strings
Previously, only generated strings had their bounds checked when a level
was parsed. This has now been fixed; Open Manifold will now tell you if
a level string is too long and truncate or lengthen it.

This also adds numbers to the error messages to make it more clear to
users what sequence specifically is too long, and in debug mode, both
the generated and user sequences are printed to make this even more clear.