Generalize music-loading functions
This change removes the separate-yet-nearly-identical functions for
loading sandbox and menu music, and replaces them with a generic function
to load a given filename from `assets/music`.

This commit also adds console log messages when loading a new music file
and slightly changes the order things are loaded to create nicer-looking
log files.