Add new level-data file: tile.json
This change adds a new file that can be specified by levels, tile.json.
tile.json is similar to character.json used for character files, but for
the tile BGFX; it specifies frame data and order for tiles. It is read
from top-to-bottom much like character idle frames, and can be freely
ordered to a level author's preference.

Entry 0 has been reserved for header params and is thus ignored for now,
so those making files must add a blank/dummy array entry for now.

Levels without a tile.json, but include a tile.png, will still work thanks
to a fallback function that's also been implemented. This fallback should
look identical to the old behavior of tile.

Below is an example of a tile.json file:
[
    {"this_space": "intentionally_blank"},
    {"x":0,  "y":0,"w":16,"h":16},
    {"x":16,  "y":0,"w":16,"h":16},
    {"x":32,  "y":0,"w":16,"h":16},
    {"x":48,  "y":0,"w":16,"h":16},
    {"x":64,  "y":0,"w":16,"h":16},
    {"x":80,  "y":0,"w":16,"h":16},
    {"x":96,  "y":0,"w":16,"h":16},
    {"x":112,  "y":0,"w":16,"h":16}
]

Fixes #18.