[BGFX] Add speed parameter to tile.json
This change allows a tile.json to tell Open Manifold how quickly
it should be cycling through frames. The speed parameter takes an
integer in milliseconds ("speed": 1000 = 1 second for example).

If speed is set to zero, the speed will be set to one beat.

Future parameters that take integers will likely require a new
struct to replace tile_frames, or additional aux_ints.