Add reloading functions to level select
Whenever the game throws up an error screen while trying to parse
a level, or if the levels folder isn't found, the player can now
press Cross to attempt to reload after fixing the problem.

In the case of level folders going missing, this will also
re-scan the level folder. Maybe this could be added as its
own button down the line, who knows?