[BGFX] Wrap `effect_id` parameters into functions
This commit changes quite a bit, but essentially all it's doing
is moving from passing effect_id as a parameter into a function
that gets it. Not only does this cut down on arguments, but it
will allow us to move background effects into their own file
down the road; something I've been wanting to do for some time.

As a side effect, this commit also makes the "Initializing BGFX"
message actually print the name of the effect instead of the
internal enum value *and* moves the conversion function into
graphics.cpp, next to all of the background effects themselves.
Handy!