[BGFX] Add new effect: `bigbang`
A strange effect that came about when I was trying to implement
draw_shape_outline() that backfired spectacularly. It's a ton
of lines that slowly circle around towards the center, eventually
forming a rounded square before exploding into a moiré effect
and then fading back away in a cycle. Very hypnotic to watch.