Add high scores and `hiscores.json`
Not *quite* as detailed as my original proposal, but it gets
the job done for now. Hopefully this can be improved later down
the road to add more metadata like playcounts and whether
it's been cleared (I'll likely change the filename when that
happens so we can migrate data from one to the other).
It's totally possible, I'm sure of it, just not particularly
high-priority at this point for me. :<

Right now we only save scores if a level is cleared, so the
mere presence of a non-zero score right now doubles as a clear
marker. This is probably subject to change in the near-future,
so I'd recommend keeping your scores backed up or recorded!

Solves #30.