Implement language files
This commit switches almost every string shown to the player from a
hardcoded one to one fetched from a giant, internal JSON defined in
main.cpp.

Most functions that used arrays of strings, like sandbox or tutorial,
now instead carry identifiers for this JSON. You can fetch any ID from
this JSON using the function "get_lang_string(string)".

Finally, an example language file is now included. Currently we hard-code
load this file, as well as a duplicate used as a fallback internally.
The two should ideally be kept 1:1 when possible.

One unfortunate downside is that, at least for me, my C++ compiler doesn't
warn or even see an error if the JSON object is not properly defined.
If you compile this and then get met with an error right away, that means
something's gone wrong with this internal language JSON.